display warning for iOS AdMob problem
If one tries to display an AdMob banner ad on iOS, before the ad has finished caching, then it will pretend to show the ad and post to /adview, but not actually show the ad. The solution is to wait until a banner ad has been cached, and then call ShowBanner. Log a warning if someone tries to do otherwise.